+++ src/main/java/com/kruemblegard/CommonModEvents.java
@@ -4,9 +4,18 @@ import com.kruemblegard.entity.KruemblegardBossEntity;
 import com.kruemblegard.entity.TraprockEntity;
 import com.kruemblegard.registry.ModEntities;
 
+import net.minecraft.core.BlockPos;
+import net.minecraft.core.Direction;
+import net.minecraft.util.RandomSource;
+import net.minecraft.world.entity.MobSpawnType;
+import net.minecraft.world.entity.SpawnPlacements;
+import net.minecraft.world.entity.monster.Monster;
 import net.minecraft.world.entity.monster.Silverfish;
+import net.minecraft.world.level.ServerLevelAccessor;
+import net.minecraft.world.level.levelgen.Heightmap;
 
 import net.minecraftforge.event.entity.EntityAttributeCreationEvent;
+import net.minecraftforge.event.entity.SpawnPlacementRegisterEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
@@ -19,5 +28,39 @@ public final class CommonModEvents {
         event.put(ModEntities.KRUEMBLEGARD.get(), KruemblegardBossEntity.createAttributes().build());
         event.put(ModEntities.TRAPROCK.get(), TraprockEntity.createAttributes().build());
         event.put(ModEntities.PEBBLIT.get(), Silverfish.createAttributes().build());
+    }
+
+    @SubscribeEvent
+    public static void onSpawnPlacementRegister(SpawnPlacementRegisterEvent event) {
+        event.register(
+                ModEntities.TRAPROCK.get(),
+                SpawnPlacements.Type.ON_GROUND,
+                Heightmap.Types.MOTION_BLOCKING_NO_LEAVES,
+                CommonModEvents::canSpawnOnSolidGround,
+                SpawnPlacementRegisterEvent.Operation.REPLACE
+        );
+
+        event.register(
+                ModEntities.PEBBLIT.get(),
+                SpawnPlacements.Type.ON_GROUND,
+                Heightmap.Types.MOTION_BLOCKING_NO_LEAVES,
+                CommonModEvents::canSpawnOnSolidGround,
+                SpawnPlacementRegisterEvent.Operation.REPLACE
+        );
+    }
+
+    private static boolean canSpawnOnSolidGround(
+            net.minecraft.world.entity.EntityType<? extends Monster> type,
+            ServerLevelAccessor level,
+            MobSpawnType spawnType,
+            BlockPos pos,
+            RandomSource random
+    ) {
+        BlockPos below = pos.below();
+        if (!level.getBlockState(below).isFaceSturdy(level, below, Direction.UP)) {
+            return false;
+        }
+
+        return Monster.checkMonsterSpawnRules(type, level, spawnType, pos, random);
     }
 }